32 bit with no errors

+++ test/training.jl
@@ -15,7 +15,7 @@ function import_ani_data(filename::String)
     # data vectors to return
     symbols = Vector{Vector{String}}()
     coordinates = Vector{Matrix{Float32}}()
-    energies = Vector{Float64}()
+    energies = Vector{Float32}()
 
     # extract all data
     for data in adl
